Refactoring extension to support blueprints
As the decorator exposed by this extension needs to be used in any route that should be protected this can include blueprints.

This presents a problem when the application is created as a factory as we can’t easily pass objects such as ‘app.auth’ into blueprints outside of a request context.

As we need to decorate routes, which is outside such a context, we need to define the ‘auth’ object outside of the application, however we still need to configure this object with information from the app.

Way to do this is to delcare the auth object outside of the application factory, but call a method it exposes to then configure it inside.

This externally defined object can then be imported anywhere it is needed, including blueprints.

At the same time I have simplified the extension to accept an application object, from which the relevant configuration settings will be read. I have also removed the ability to set a JWKS since this is only needed in testing.

To ensure tests don’t break, I have further adapted the test version of the extension to provide the various config settings as before (including the JWKS), as this doesn’t need to support blueprints).