Adding support for testing tokens in other applications
Currently all of the functionality to generate test tokens, signed by test signing keys, is internal to this package for it’s integration tests - as we have do things like break tokens and keys.

In applications that use this provider, they need to test routes protected by this provider, which either requires them to request a geniue token (which will require credentials, and for methods requring scopes, permissions assigned, which may require extensive sign-off) or to also generate test tokens and keys.

It’s assumed that most applications will want to take the later approach, so support for issuing tokens and keys should be added to this project, for use by implementing applications, rather than being duplicated in each of them.

To that end, the existing TestJwk and TestJwt classes have been refactored to expose the elements needed to create realistic tokens and keys within the main package.

Some functionality, such as breaking tokens/keys is retained internally, as it’s not expected that this should be needed in implementating applications, and may give the impression we aren’t implementing the standards properly.