Allowing validators to be reregistered
This is needed to allow the token validator to be replaced with a new validator instance.

This is needed when using this provider in a testing environment, as token validator instances contain a local JWK which is reset on each test. Typically however the ‘auth’ object is instantiated outside of an test’s environment, and so persists between them, but the JWK stored in the application (and used to generate test tokens) is inside this environment, and so creates a mis-match.

This new method allows an application environment to effectively bind the validator instance to itself, preventing instances leaking into other tests.