Removing test objects from main code
The TestJwt and TestJwks classes are used for generating fake JWTs and JWKSs during testing (either directly within this package or when this package is used in other applications).

These classes are needed to avoid making calls to Azure, either because this is undesirable when testing or to test failure modes that are not common, or possible, in Azure.

Previously test classes where used as part of main application, enabled when Flask’s test mode was enabled. This was not very clean and quite inflexible.

Now test classes have been moved to a `mocks` module (as they effectively mock the main JWT and JWKS classes) and are used by patching the relevant methods in the main package to use test classes instead.

Tests for this package and the README testing example have been updated to use this new structure/behaviour. Users of this package will need to update their own tests accordingly as well, making this a breaking change.