Adding JWKS error
Requires setting up our own JWK & JWKS for testing. As these are fetched from a OIDC endpoint we also need to fake that using a testing web-server.

As the JWKS is internal or server side, exceptions need to support different status codes (i.e. not just 401, 500 in this case). Also namespacing exceptions consistently.